refactoring into subroutines vs main
